refactor(mcap): extract temporal tag logic into useMcapTemporalTags hook
